Change the name of the documentation entry for PCA-related functions
Using ***.pca was problematic in case a package defined objects of class pca. 

In addition, it makes sense for autoplot_pca to be exported since this is where some of the work at least, is carried out.

But this is still mostly tricking roxygen2, there must be a better way