In autoplot, pass ... to geoms rather than to fortify
This allows to set aesthetics in addition to mapping them through the mapping argument.
This requires, however, that autoplot has all the arguments of fortify, to be able to pass them on. In particular, this adds `data` and `PC`.
Passing data requires a default NULL value and therefore the extraction of data from the object is now done within each function rather than in the arguments.